add script to make Makefile

+++ include/os_spec/kfs.h
@@ -3,6 +3,7 @@
 
 #include <stdbool.h>
 #include <stddef.h>
+#include <stdint.h>
 #include <stdio.h>
 
 /* Due to platform-specificity, this file defines some wrapper functions
@@ -12,6 +13,6 @@ char* kfs_get_bin_dir(void);
 char* kfs_get_relpath(const char* from, const char* to);
 char* kfs_trunc_leaf(const char* path);
 bool kfs_is_relative(const char* path);
-size_t kfs_file_size(FILE* file);
+uintmax_t kfs_file_size(FILE* file);
 
 #endif
